FraudBackend/app/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database and graph
neo4j==5.14.1
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Background tasks and async
celery==5.3.4
redis==5.0.1
asyncio-mqtt==0.16.1

# HTTP client for external APIs
httpx==0.25.2

# Data processing and utilities
pandas==2.1.4
# Use a Python 3.12-compatible wheel to avoid source build/distutils errors.
numpy==1.26.4
python-dateutil==2.8.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and configuration
python-dotenv==1.0.0

# LDAP / AD
ldap3==2.9.1
